fix(css): improve built-in support with css prop as function (#181)
* chore: refactor getGlamorClassName

* fix(css): improve built-in support with css prop as function

**What**: This refactors things a bit and separates `css` the prop from
the overrides props.

**Why**: Otherwise, if the `css` prop happens to be a function, the
`propsAreCssOverrides` would make the props be assigned to the function
and then they'd never actually be applied. By splitting them from each
other it allows us to merge both of them.

**How**: Most of the work happened in `get-glamor-classname.js`. I first
refactored it to simplify things a bit. Then I plugged in the `cssProp`
and `cssOverrides` into their order of precidence. I also added a bit of
recursion to the `handleStyles` function which allows you to pass an
array of functions, this is useful for the `css` prop in particular. So
you can compose functions together without a `compose` helper.